nutrition-ai-agent/requirements.txt
fastapi
uvicorn

langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-groq
groq

faiss-cpu
sentence-transformers

python-dotenv
pydantic
pymongo>=4.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.2.0
pytest
pypdf>=5.0.0
fpdf2>=2.8.4
python-multipart>=0.0.9